refactored common layers
