--- src/pytorch_tabular/models/common/utils.py
@@ -1,24 +1,0 @@
-import torch
-
-
-def to_one_hot(y, depth=None):
-    r"""Takes integer with n dims and converts it to 1-hot representation with n + 1 dims.
-
-    The n+1'st dimension will have zeros everywhere but at y'th index, where it will be equal to 1.
-    Args:
-        y: input integer (IntTensor, LongTensor or Variable) of any shape
-        depth (int):  the size of the one hot dimension
-    """
-    y_flat = y.to(torch.int64).view(-1, 1)
-    depth = depth if depth is not None else int(torch.max(y_flat)) + 1
-    y_one_hot = torch.zeros(y_flat.size()[0], depth, device=y.device).scatter_(1, y_flat, 1)
-    y_one_hot = y_one_hot.view(*(tuple(y.shape) + (-1,)))
-    return y_one_hot
-
-
-def _make_ix_like(input, dim=0):
-    d = input.size(dim)
-    rho = torch.arange(1, d + 1, device=input.device, dtype=input.dtype)
-    view = [1] * input.dim()
-    view[0] = -1
-    return rho.view(view).transpose(0, dim)
