+++ src/pytorch_tabular/models/gate/components.py
@@ -7,7 +7,7 @@ from typing import Callable
 import torch
 import torch.nn as nn
 
-from pytorch_tabular.models.common.activations import entmax15, RSoftmax
+from pytorch_tabular.models.common.layers.activations import entmax15, RSoftmax
 
 
 class NeuralDecisionStump(nn.Module):
